+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix `Map.prototype.insertOrUpdateComputed()` to bind `this` to the map
+
 ## [v0.6.0] - 2026-01-06
 
 ### Changed